linux: strip so

+++ bin/deploy.sh
@@ -9,6 +9,8 @@ log() {
 log "Ensuring artifact folders are present..."
 test -e artifacts/linux-x86-64 || exit 2
 test -e artifacts/macos-x86-64 || exit 2
+test -e artifacts/win32-i386 || exit 2
+test -e artifacts/win32-x86-64 || exit 2
 
 log "Copying artifacts into their respective packages..."
 cp artifacts/linux-x86-64/lib/x86_64-linux-gnu/libargon2.so.1 libargon2-x86_64-linux/libargon2.so
